+++ src/rlfmi.rs
@@ -1,8 +1,8 @@
+use crate::backend::SearchIndexBackend;
 use crate::character::{prepare_text, Character};
 #[cfg(doc)]
 use crate::converter;
 use crate::converter::{Converter, IndexWithConverter};
-use crate::iter::SearchIndexBackend;
 use crate::sais;
 use crate::search::Search;
 use crate::suffix_array::{HasPosition, SuffixOrderSampledArray};
